+++ src/main/java/org/myteam/server/news/service/NewsReadService.java
@@ -1,10 +1,14 @@
 package org.myteam.server.news.service;
 
-import org.myteam.server.news.dto.controller.response.NewsListResponse;
+import org.myteam.server.global.exception.ErrorCode;
+import org.myteam.server.global.exception.PlayHiveException;
+import org.myteam.server.news.domain.News;
+import org.myteam.server.news.dto.service.response.NewsListResponse;
 import org.myteam.server.news.repository.NewsQueryRepository;
-import org.myteam.server.news.dto.service.response.NewsDto;
+import org.myteam.server.news.dto.repository.NewsDto;
 import org.myteam.server.news.dto.service.request.NewsServiceRequest;
 import org.myteam.server.global.page.response.PageCustomResponse;
+import org.myteam.server.news.repository.NewsRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +21,7 @@ import lombok.RequiredArgsConstructor;
 public class NewsReadService {
 
 	private final NewsQueryRepository newsQueryRepository;
+	private final NewsRepository newsRepository;
 
 	public NewsListResponse findAll(NewsServiceRequest newsServiceRequest) {
 		Page<NewsDto> newsPagingList = newsQueryRepository.getNewsList(
@@ -25,7 +30,12 @@ public class NewsReadService {
 				newsServiceRequest.toPageable()
 			);
 
-		return NewsListResponse.of(PageCustomResponse.of(newsPagingList));
+		return NewsListResponse.createResponse(PageCustomResponse.of(newsPagingList));
+	}
+
+	public News findById(Long newsId) {
+		return newsRepository.findById(newsId)
+			.orElseThrow(() -> new PlayHiveException(ErrorCode.NEWS_NOT_FOUND));
 	}
 
 }
